refactor(cli): split remote install helpers
