+++ crates/surge-cli/src/commands/install/remote/types.rs
@@ -1,0 +1,114 @@
+use super::{Deserialize, Result, Serialize, SurgeError, infer_os_from_rid, logline};
+
+pub(crate) fn ensure_supported_tailscale_rid(rid: &str) -> Result<()> {
+    match infer_os_from_rid(rid) {
+        Some(os) if os == "linux" => Ok(()),
+        Some(os) => Err(SurgeError::Config(format!(
+            "Tailscale install currently supports Linux targets only. Selected RID '{rid}' targets '{os}'."
+        ))),
+        None => {
+            logline::warn(&format!(
+                "Unable to infer OS from selected RID '{rid}'. Tailscale install supports Linux targets only; verify this RID is Linux-compatible."
+            ));
+            Ok(())
+        }
+    }
+}
+
+#[derive(Debug, Clone, Default)]
+pub(crate) struct RemoteLaunchEnvironment {
+    pub(crate) display: Option<String>,
+    pub(crate) xauthority: Option<String>,
+    pub(crate) dbus_session_bus_address: Option<String>,
+    pub(crate) wayland_display: Option<String>,
+    pub(crate) xdg_runtime_dir: Option<String>,
+}
+
+impl RemoteLaunchEnvironment {
+    pub(crate) fn has_graphical_session(&self) -> bool {
+        self.display.is_some() || self.wayland_display.is_some()
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum RemoteInstallerMode {
+    Online,
+    Offline,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum RemoteTailscaleTransferStrategy {
+    AppCopy,
+    StagedInstallerCache,
+    Installer { prefer_published: bool },
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum RemoteHostInstallerAvailability {
+    Available,
+    Unavailable,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum RemoteTailscaleOperation {
+    Stage,
+    Install,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum RemoteTailscaleCachedState {
+    None,
+    AppCopyPayload,
+    InstallerCache,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum VerifiedRemoteStage {
+    AppCopyPayload,
+    InstallerCache,
+}
+
+impl VerifiedRemoteStage {
+    pub(crate) fn description(self) -> &'static str {
+        match self {
+            Self::AppCopyPayload => "staged app payload",
+            Self::InstallerCache => "staged installer cache",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) struct RemoteTailscaleTransferInputs {
+    pub(crate) host_installer_availability: RemoteHostInstallerAvailability,
+    pub(crate) installer_mode: RemoteInstallerMode,
+    pub(crate) operation: RemoteTailscaleOperation,
+    pub(crate) cached_state: RemoteTailscaleCachedState,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) struct RemoteInstallState {
+    pub(crate) version: String,
+    pub(crate) channel: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub(crate) struct RemoteStagedPayloadIdentity {
+    pub(crate) app_id: String,
+    pub(crate) version: String,
+    pub(crate) channel: String,
+    pub(crate) rid: String,
+    pub(crate) full_filename: String,
+    pub(crate) full_sha256: String,
+    pub(crate) install_directory: String,
+    pub(crate) supervisor_id: String,
+    pub(crate) storage_provider: String,
+    pub(crate) storage_bucket: String,
+    pub(crate) storage_region: String,
+    pub(crate) storage_endpoint: String,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) struct RemotePublishedInstallerPlan {
+    pub(crate) candidate_keys: Vec<String>,
+    pub(crate) blockers: Vec<String>,
+}
